Fix two uniforms that were silently never reaching the GPU
Mining the 528 run logs from the catalogue re-measurement turned up a GL error
in two scenes. Re-probed with KALEIDO_GL_DEBUG=1 it was not one error but one
PER FRAME -- 236 over a five second run of TunnelPlain, 245 for TunnelReverse.
Eighteen other scenes sampled alongside them were clean, so it was isolated.

Uniform.cpp uploads <int> and <bool> with glUniform1i and <float> with
glUniform1f. Both scenes pointed one at a GLSL uniform of the other kind:

  TunnelPlain     <float name="sides">  vs  uniform int   sides
  TunnelReverse   <bool  name="rotate"> vs  uniform float rotate

The driver rejects the call with GL_INVALID_OPERATION and the uniform keeps
its default of 0. So this was never just log noise: TunnelPlain's kaleidoscope
fold count sat at 0 and TunnelReverse's rotation could never switch on. The
scenes measurably change now that the values arrive -- TunnelPlain luma 0.343
to 0.458 and occupancy 0.99 to 0.79, TunnelReverse contrast 0.126 to 0.148.

`rotate` is only ever tested as `if (rotate > 0.0)`, so it becomes an int,
which is also how the working DarkAmbientTunnel declares the same knob.
`sides` is a fold count with an integer range, so the config declaration
becomes <int> -- in all five presets that carry the scene.

Nothing normally shows this class of bug: RenderPipeline::checkGLErrors() is a
no-op unless KALEIDO_GL_DEBUG is set, precisely because glGetError() can force
a driver sync. Worth stating plainly, since it also means the "zero GL errors"
line in the re-measurement covered less than it sounded like -- that scan ran
without the variable set.

So Tools/check_uniform_types.py now checks the whole catalogue statically:
every <float>/<int>/<bool> declaration in every config against the .frag's
actual uniform type. It found exactly these two across 641 shaders and now
reports none; it exits non-zero on a mismatch, so it can gate a build.

The baseline doc is regenerated with the two scenes re-measured -- their old
rows described a broken scene. It also gained a note on a GARISH false
positive the fix exposed: TunnelPlain now reads satHi 0.362, the 98th
percentile, purely because it samples the photo and the pinned image is a
vivid one. It invents no colour of its own, so for photo-folding scenes that
flag describes the image, not the shader.

+++ Scene2D/TunnelReverse.frag
@@ -36,7 +36,7 @@ uniform float speed;
 uniform int sides;
 uniform float power;
 uniform float speedTunnelReverse;
-uniform float rotate;
+uniform int   rotate;   // <bool> in the config -> uploaded with glUniform1i
 uniform float audioPhase;     // integrated audio rotation phase (radians, jump-free)
 uniform float audioAdvance;   // integrated audio tunnel advance (jump-free)
 uniform float audioKick;      // subtle brightness pulse on kicks
@@ -64,7 +64,7 @@ void main() {
     
     
     float angle = 0.0; 
-    if( rotate > 0.0 )
+    if( rotate > 0 )
        angle = M_PI*0.25;
     
     // a rotation
